Modularise the model selection subwidget
